Update App.jsx
Changed the initial state of the isLoggedIn to be: const [isLoggedIn, setIsLoggedIn] = useState(!!localStorage.getItem("token"));
9:54
This way the state never stars as false, but instead it will check if the token is there or not every time the page refreshes.
This solved the issue of going back to the login page when the page was refreshing.